SPRY-3032 Rename CacheManagerGui namespace to CacheManager in CacheManager model and related classes.
