+++ src/ValanticSpryker/Zed/CacheManager/Business/CacheManagerFacadeInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types = 1);
+
+namespace ValanticSpryker\Zed\CacheManager\Business;
+
+interface CacheManagerFacadeInterface
+{
+
+}
